# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Киномания",
        "version": "1.0"
    },
    "paths": {
...
        "/api/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Get favorite films for the authenticated user",
                "operationId": "81123175f76259d611e2c337477dfa5f",
                "responses": {
                    "200": {
                        "description": "List of favorite films"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/favorites/{filmId}/{status}": {
            "patch": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Update favorite status of a film",
                "operationId": "88d32e1b445662df285689382fd48203",
                "parameters": [
                    {
                        "name": "filmId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Film not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/films": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Get a list of films",
                "operationId": "271e474415d2f29bf3c163247c5f4639",
                "responses": {
                    "200": {
                        "description": "List of films"
                    }
                }
            },
            "post": {
                "tags": [
                    "Films"
                ],
                "summary": "Add a new film",
                "operationId": "ce1501bac824524d1e4706075650bf11",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "imdb"
                                ],
                                "properties": {
                                    "imdb": {
                                        "type": "string",
                                        "example": "tt1234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Film added to moderation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Film added to moderation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/films/{film}": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Get a specific film by ID",
                "operationId": "5f8edfd1839b95f178d8fed97762e4e6",
                "parameters": [
                    {
                        "name": "film",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Film details"
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Films"
                ],
                "summary": "Update a specific film",
                "operationId": "58fecd7f62d641cf7267e28af626b186",
                "parameters": [
                    {
                        "name": "film",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Film updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Film updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            }
        },
        "/api/films/{film}/similar": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Get similar films",
                "operationId": "6dd8c792b4d25ba9f43143edc238b637",
                "parameters": [
                    {
                        "name": "film",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of similar films"
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            }
        },
...
        "/api/genres/{genre}": {
            "put": {
                "tags": [
                    "Genres"
                ],
                "summary": "Update a specific genre",
                "operationId": "ec8267c8b015ed7fae17a69b06435c3f",
                "parameters": [
                    {
                        "name": "genre",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Action"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Genre updated successfully"
                    },
                    "404": {
                        "description": "Genre not found"
                    }
                }
            }
        },
...
    },
...
}